WIP scooping interface
