+++ modules/luigi/payment_server_mockup.py
@@ -10,5 +10,5 @@ def get_amount_inserted(req):
 
 if __name__ == "__main__":
     rospy.init_node('payment_mockup_server')
-    rospy.Service('payment_mockup', Payment, get_amount_inserted)
+    rospy.Service('payment', Payment, get_amount_inserted)
     rospy.spin()
